test: rename to PDFDocumentHandlerTest, table-driven cases, fixture in Materials
Addressing PR feedback:

1) Rename WriteObjectRecursionCap.cpp -> PDFDocumentHandlerTest.cpp to
   match the <ProductionFile>Test.cpp convention (the bug is in
   PDFDocumentHandler.cpp). Entry function renamed accordingly per
   CMake's create_test_sourcelist requirement.

2) Restructure internals to the sc<Function>Cases / Run<Function>Cases
   table-driven pattern used elsewhere (see CFFFileInputTest, PR #364).
   Table: scWriteObjectByTypeCases. Runner: RunWriteObjectByTypeCases.
   Case labels are <Condition>_<Result> only — function context lives
   on the runner and in the failure-message prefix.

3) Source PDF is now a fixture in Materials/EmptyOnePage.pdf (480
   bytes, produced once by PDFWriter and committed as binary per
   .gitattributes). Previously the test regenerated it on every run,
   which was redundant.

No behavior change in the fix itself (the cap and threading in
PDFDocumentHandler.{h,cpp} are unchanged). All 102 tests pass.
