+++ PDFWriterTesting/Materials/EmptyOnePage.pdf
@@ -1,0 +1,39 @@
+%PDF-1.4
+%½¾¼µ
+1 0 obj
+<<
+	/Type /Page
+	/Parent 2 0 R
+	/MediaBox [ 0 0 100 100 ]
+	/Resources <<
+	>>
+>>
+endobj
+2 0 obj
+<<
+	/Type /Pages
+	/Count 1
+	/Kids [ 1 0 R ]
+>>
+endobj
+3 0 obj
+<<
+	/Type /Catalog
+	/Pages 2 0 R
+>>
+endobj
+xref
+0 4
+0000000000 65535 f
+0000000017 00000 n
+0000000121 00000 n
+0000000190 00000 n
+trailer
+<<
+	/Size 4
+	/Root 3 0 R
+	/ID [ <BFCF236245A20616F0B2D715A588223B> <BFCF236245A20616F0B2D715A588223B> ]
+>>
+startxref
+247
+%%EOF
